requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
Pillow>=10.0
python-multipart>=0.0.9
Jinja2>=3.1.0
aiofiles>=24.0
